+++ app/storybook/.storybook/manager.ts
@@ -2,28 +2,8 @@ import { addons } from '@storybook/manager-api'
 import { defaultTheme } from './defaultTheme'
 
 const env = process.env.NODE_ENV
-const devOnly = [
-  'Components/PieChart',
-  'Hooks/useTopValues',
-  'Hooks/useDataGrid',
-  'API/DataGridQueryProps',
-  'API/PieChartQueryProps',
-  'API/TopValuesQueryProps',
-  'API/RecordsByIdQueryProps',
-  'Hooks/useRecordsById',
-  'Getting started/Query Metrics',
-  'API/AverageMetricQueryInput',
-  'API/CountDistinctMetricQueryInput',
-  'API/CounterQueryProps',
-  'API/CustomMetricQueryInput',
-  'API/DataPoolInput',
-  'API/MaxMetricQueryInput',
-  'API/MetricInput',
-  'API/MinMetricQueryInput',
-  'API/SumMetricQueryInput',
-  'Hooks/useSql',
-  'API/SqlQueryProps'
-]
+const devOnly = ['']
+const hidden = ['Components/Autocomplete']
 
 addons.setConfig({
   theme: defaultTheme,
@@ -34,7 +14,11 @@ addons.setConfig({
       },
       devOnly: (item) => {
         return (
-          env !== 'production' || !devOnly.map((element) => element.toLowerCase()).includes(item.title.toLowerCase())
+          env !== 'production' ||
+          !devOnly
+            .concat(hidden)
+            .map((element) => element.toLowerCase())
+            .includes(item.title.toLowerCase())
         )
       }
     }