+++ packages/ui-kit/src/hooks/useCounter/useCounter.ts
@@ -1,4 +1,4 @@
-import { CounterQueryProps, useAccessToken, useLog } from '../../components'
+import { CounterQueryProps, useAccessToken, useFilters, useLog } from '../../components'
 import { CounterQuery, getTimeZone, PROPEL_GRAPHQL_API_ENDPOINT, useCounterQuery } from '../../helpers'
 import { UseQueryProps } from '../types/Query.types'
 
@@ -15,9 +15,10 @@ export const useCounter = (props: CounterQueryProps): UseQueryProps<CounterQuery
     propelApiUrl,
     metric,
     timeRange,
-    filters,
+    filters: filtersFromProp,
     refetchInterval,
     retry,
+    enabled: enabledProp = true,
     timeZone
   } = props
 
@@ -33,7 +34,11 @@ export const useCounter = (props: CounterQueryProps): UseQueryProps<CounterQuery
   // Get access token first from props, then if it is not provided via prop get it from provider
   const accessToken = accessTokenFromProp ?? accessTokenFromProvider
 
-  const enabled = accessToken != null
+  const { filters: filtersFromProvider } = useFilters()
+
+  const filters = filtersFromProp ?? filtersFromProvider
+
+  const enabled = accessToken != null && enabledProp
 
   // Log error if no access token provided and metric is provided
   if (!enabled && metric) {
@@ -68,7 +73,7 @@ export const useCounter = (props: CounterQueryProps): UseQueryProps<CounterQuery
           start: timeRange?.start ?? null,
           stop: timeRange?.stop ?? null
         },
-        filters: filters ?? []
+        filters
       }
     },
     {